% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_lead.R
\docType{data}
\name{comp}
\alias{comp}
\title{Leader Data Set}
\format{A data frame with 24 rows and 4 variables:
\describe{
  \item{company}{individual ID}
  \item{profit}{if the company had a profit}
  \item{size}{relative size of the company}
  \item{ceo}{ceo}
}}
\source{
\url{https://www.afhayes.com/regression-analysis-and-linear-models.html}
}
\usage{
comp
}
\description{
A small data set with n = 24 individuals about for-profit companies.
}
\keyword{datasets}
