% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cond_effect.R
\name{cond_effect}
\alias{cond_effect}
\title{Conditional Effects in Interactions}
\usage{
cond_effect(data, formula, where = NULL)
}
\arguments{
\item{data}{the data frame}

\item{formula}{the formula that works with \code{lm()}}

\item{where}{a two-sided formula describing a which and where we should do the conditional effect (e.g., \code{x1 ~ x1 - 5} for centering x1 at 5)}
}
\value{
an lm object
}
\description{
Obtain the inferential statistics for a conditional effect
}
