% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{diagnostics}
\alias{diagnostics}
\title{Diagnostic Measures for a Linear Regression}
\usage{
diagnostics(model)
}
\arguments{
\item{model}{The linear model object (\code{lm()} or \code{glm()})}
}
\description{
This function is a wrapper of several diagnostic functions and combines these measures
into a single tidy data.frame.
}
