% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_exer.R
\docType{data}
\name{exer}
\alias{exer}
\title{Exercise Data Set}
\format{A data frame with 10 rows and 6 variables:
\describe{
  \item{id}{individual ID}
  \item{exercise}{level of exercise (grouping variable with 0, 2, and 4)}
  \item{food}{food}
  \item{metab}{metab}
  \item{sex}{sex of the individual (0 = male, 1 = female)}
  \item{wtloss}{how much weight was lost}
}}
\source{
\url{https://www.afhayes.com/regression-analysis-and-linear-models.html}
}
\usage{
exer
}
\description{
A small data set with n = 10 individuals about exercise and weight loss.
}
\keyword{datasets}
