% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_olr.R
\name{glm_olr}
\alias{glm_olr}
\title{Diagnostic Measures for a Linear Regression}
\usage{
glm_olr(formula, data, ...)
}
\arguments{
\item{formula}{a two-sided regression formula with a single outcome (a 3+ level factor) and predictors}

\item{data}{the data frame}

\item{...}{arguments passed to \code{MASS::polr()}}
}
\description{
This function is a wrapper of several diagnostic functions and combines these measures
into a single tidy data.frame.
}
