% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_golf.R
\docType{data}
\name{golf}
\alias{golf}
\title{Golf Data Set}
\format{A data frame with 24 rows and 3 variables:
\describe{
  \item{id}{individual ID}
  \item{plays}{grouping variable with 0, 2, and 4)}
  \item{points}{points}
}}
\source{
\url{https://www.afhayes.com/regression-analysis-and-linear-models.html}
}
\usage{
golf
}
\description{
A dataset containing n = 24 individuals about golf
}
\keyword{datasets}
