% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_hosp.R
\docType{data}
\name{hosp}
\alias{hosp}
\title{Hospital Data Set}
\format{A data frame with 300 rows and 9 variables:
\describe{
  \item{exhaust}{exhaust}
  \item{safety}{safety}
  \item{injury}{injury}
  \item{tenure}{tenure}
  \item{serious}{serious}
  \item{injuryb}{injuryb}
  \item{latency}{latency}
  \item{sex}{sex of the individual (0 = male, 1 = female)}
  \item{age}{age of the individual}
}}
\source{
\url{https://www.afhayes.com/regression-analysis-and-linear-models.html}
}
\usage{
hosp
}
\description{
A small data set with n = 300 individuals about a hospital.
}
\keyword{datasets}
