% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_married.R
\docType{data}
\name{marr}
\alias{marr}
\title{Married Data Set}
\format{A data frame with 20 rows and 4 variables:
\describe{
  \item{mstatus}{marriage status}
  \item{satis}{satisfaction level}
  \item{income}{relative size of the company}
  \item{sex}{sex of the individual}
}}
\source{
\url{https://www.afhayes.com/regression-analysis-and-linear-models.html}
}
\usage{
marr
}
\description{
A small data set with n = 24 individuals about for-profit companies.
}
\keyword{datasets}
