% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mt_adjust.R
\name{mt_adjust}
\alias{mt_adjust}
\title{Multiple Testing Adjustment for Regression}
\usage{
mt_adjust(model, adjustment)
}
\arguments{
\item{model}{the regression model}

\item{adjustment}{the type of adjustment}
}
\value{
a data.frame
}
\description{
Obtain the regression results with a multiple testing adjustment
}
