% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oddsratios.R
\name{oddsratios}
\alias{oddsratios}
\title{Odds Ratios}
\usage{
oddsratios(model)
}
\arguments{
\item{model}{The generalized linear model object (from \code{glm()})}
}
\description{
This function provides the exponentiated coefficients and confidence intervals
from a logistic regression model.
}
