% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_poli.R
\docType{data}
\name{poli}
\alias{poli}
\title{Politics Data Set}
\format{A data frame with 340 rows and 16 variables:
\describe{
  \item{pknow}{political knowledge}
  \item{age}{age of the individual}
  \item{educ}{educational level of the individual}
  \item{sex}{sex of the individual}
  \item{income}{income (in thousands)}
  \item{polint}{political interest}
  \item{party}{political party of the individual}
  \item{libcon}{liberal/conversative rating}
  \item{pdiscuss}{political discussions}
  \item{natnews}{natnews}
  \item{npnews}{news consumption}
  \item{locnews}{news consumption at local level}
  \item{talkrad}{news consumption from talk radio}
  \item{ses}{Socio-Economic Status of the individual}
  \item{news}{news consumption}
  \item{demoneg}{demoneg}
}}
\source{
\url{https://www.afhayes.com/regression-analysis-and-linear-models.html}
}
\usage{
poli
}
\description{
A small data set with n = 24 individuals about for-profit companies.
}
\keyword{datasets}
