% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_poverty.R
\docType{data}
\name{poverty}
\alias{poverty}
\title{Poverty Data Set}
\format{A data frame with 125 rows and 26 variables:
\describe{
  \item{State}{State name}
  \item{poverty_pct}{the poverty rate for the state}
  \item{birthrate_15-17}{Birth rate for females aged 15 - 17 years old in the state}
  \item{birthrate_18-19}{Birth rate for females aged 18 - 19 years old in the state}
  \item{ViolentCrime}{The violent crime rate for the state}
  \item{TeenBirth}{The overall teen birth rate for the state}
}}
\usage{
poverty
}
\description{
A small data set with n = 51 states (and the District of Columbia) regarding poverty rates,
teen birth rates, and violent crime rate.
}
\keyword{datasets}
