% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/oddsratios.R
\name{riskratios}
\alias{riskratios}
\title{Risk Ratios}
\usage{
riskratios(model)
}
\arguments{
\item{model}{The generalized linear model object (from \code{glm()})}
}
\description{
This function provides the exponentiated coefficients and confidence intervals
from a logistic regression model.
}
